requirements.txt
Flask==3.0.3
requests~=2.32.0
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1
pypdf2~=3.0.1
pandas~=2.2.2
python-docx~=1.1.2